use partial, uncached chunks on large offsets
a request that starts in the middle of a chunk will have a great delay,
because no useful data will be sent before the chunk is cached up to
that offset.

to get rid of this delay, we sacrifice caching of the first chunk in a
request, if the offset is more than 20% of that chunk.